refactor: decompose EntityStore into specialized services and DI
